+++ packages/angular-workspace/ok-angular/fv/search-input/public-api.ts
@@ -1,0 +1,2 @@
+export * from './ok-fv-search-input.directive';
+export * from './ok-fv-search-input.module';