+++ packages/react-workspace/nimble-react/src/anchor-tab/index.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { AnchorTab, anchorTabTag } from '@ni/nimble-components/dist/esm/anchor-tab';
 import { wrap } from '../utilities/react-wrapper';