+++ packages/react-workspace/nimble-react/src/breadcrumb-item/index.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { BreadcrumbItem, breadcrumbItemTag } from '@ni/nimble-components/dist/esm/breadcrumb-item';
 import { wrap } from '../utilities/react-wrapper';