+++ packages/react-workspace/nimble-react/src/step/index.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { Step, stepTag } from '@ni/nimble-components/dist/esm/step';
 import { wrap } from '../utilities/react-wrapper';