+++ packages/react-workspace/nimble-react/src/table-column/text/index.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { TableColumnText, tableColumnTextTag } from '@ni/nimble-components/dist/esm/table-column/text';
 import { wrap } from '../../utilities/react-wrapper';