+++ packages/react-workspace/nimble-react/src/theme-provider/index.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { ThemeProvider, themeProviderTag } from '@ni/nimble-components/dist/esm/theme-provider';
 import { Theme } from '@ni/nimble-components/dist/esm/theme-provider/types';
 import { wrap } from '../utilities/react-wrapper';