+++ packages/angular-workspace/ok-angular/CHANGELOG.md
@@ -1,8 +1,25 @@
 # Change Log - @ni/ok-angular
 
-<!-- This log was last generated on Tue, 21 Apr 2026 19:47:00 GMT and should not be manually modified. -->
+<!-- This log was last generated on Thu, 30 Apr 2026 22:40:17 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 2.2.1
+
+Thu, 30 Apr 2026 22:40:17 GMT
+
+### Patches
+
+- Bump @ni/ok-components to v1.1.1
+
+## 2.2.0
+
+Tue, 28 Apr 2026 14:27:06 GMT
+
+### Minor changes
+
+- feat: add Ok Fv Search Input ([ni/nimble@eb9e67b](https://github.com/ni/nimble/commit/eb9e67b30b075f0f53a39d69fe39bf866b9cf81f))
+- Bump @ni/ok-components to v1.1.0
 
 ## 2.1.6